src/pytorch_requirements.txt
argon2-cffi @ file:///tmp/build/80754af9/argon2-cffi_1596828493937/work
async-generator==1.10
attrs @ file:///tmp/build/80754af9/attrs_1600298409949/work
Automat==22.10.0
backcall==0.2.0
bleach @ file:///tmp/build/80754af9/bleach_1600439572647/work
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1602537222527/work
chardet==3.0.4
cmake==4.0.2
constantly==15.1.0
cryptography==41.0.3
cssselect==1.2.0
cycler==0.10.0
Cython==0.29.21
dataclasses==0.6
deap==1.3.1
decorator==4.4.2
defusedxml==0.6.0
distro==1.8.0
easyocr==1.6.2
entrypoints==0.3
filelock==3.12.2
flowkey-dl==0.1.4
fpdf==1.7.2
fsspec==2025.3.0
future==0.18.2
hyperlink==21.0.0
idna==2.10
imageio @ file:///tmp/build/80754af9/imageio_1594161405741/work
imageio-ffmpeg==0.4.2
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
incremental==22.10.0
ipykernel @ file:///tmp/build/80754af9/ipykernel_1596207638929/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///tmp/build/80754af9/ipython_1598883821407/work
ipython_genutils==0.2.0
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1601490159889/work
itemadapter==0.8.0
itemloaders==1.1.0
jedi @ file:///tmp/build/80754af9/jedi_1598371611696/work
Jinja2==2.11.2
jmespath==1.0.1
joblib @ file:///tmp/build/80754af9/joblib_1601912903842/work
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter==1.0.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1598884538475/work
jupyter-core==4.6.3
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1604014535162/work
labelImg==1.8.4
lit==18.1.8
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
mpmath==1.3.0
nb-conda==2.2.1
nb-conda-kernels @ file:///tmp/build/80754af9/nb_conda_kernels_1606775941989/work
nbclient @ file:///tmp/build/80754af9/nbclient_1602783176460/work
nbconvert @ file:///tmp/build/80754af9/nbconvert_1601914830498/work
nbformat @ file:///tmp/build/80754af9/nbformat_1602783287752/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1601499549014/work
networkx==2.8.7
ninja==1.10.2.4
notebook @ file:///tmp/build/80754af9/notebook_1601501575118/work
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603570489231/work
numpy-stl==2.13.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
olefile==0.46
opencv-python-headless==4.5.4.60
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas @ file:///tmp/build/80754af9/pandas_1602088120436/work
pandocfilters==1.4.2
parsel==1.8.1
parso==0.7.0
pexpect==4.8.0
pickle5==0.0.11
pickleshare==0.7.5
Pillow @ file:///tmp/build/80754af9/pillow_1603822255246/work
prometheus-client==0.8.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
Protego==0.3.0
psutil @ file:///tmp/build/80754af9/psutil_1612298023621/work
ptflops==0.6.7
ptyprocess==0.6.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
PyDispatcher==2.0.7
Pygments @ file:///tmp/build/80754af9/pygments_1600458456400/work
pyOpenSSL==23.2.0
pyparsing==2.4.7
PyPika==0.48.9
PyQt5==5.12.3
PyQt5_sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyrsistent @ file:///tmp/build/80754af9/pyrsistent_1600141720057/work
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.1
python-utils==2.4.0
pytz==2020.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==19.0.2
qtconsole @ file:///tmp/build/80754af9/qtconsole_1600870028330/work
QtPy==1.9.0
queuelib==1.6.2
requests==2.25.0
requests-file==1.5.1
scikit-build==0.15.0
scikit-image==0.19.3
scikit-learn @ file:///tmp/build/80754af9/scikit-learn_1598376899566/work
scipy @ file:///tmp/build/80754af9/scipy_1597686649129/work
Scrapy==2.10.0
seaborn @ file:///tmp/build/80754af9/seaborn_1600553570093/work
Send2Trash==1.5.0
service-identity==23.1.0
Shapely==1.8.4
sip==4.19.13
six==1.15.0
sklearn-genetic==0.3.0
sympy==1.13.3
terminado==0.9.1
testpath==0.4.4
thop==0.0.31.post2005241907
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
tifffile==2022.8.12
tldextract==3.4.4
torch==2.0.1+cu117
torch-summary==1.4.3
torchaudio==0.7.0a0+ac17b64
torchvision==0.15.2+cu117
tornado==6.0.4
tqdm @ file:///tmp/build/80754af9/tqdm_1602185206534/work
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
triton==2.0.0
Twisted==22.10.0
typing_extensions==4.13.2
urllib3==1.26.2
w3lib==2.1.2
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
widgetsnbextension==3.5.1
ZConfig==3.5.0
zdaemon==4.3
zipp @ file:///tmp/build/80754af9/zipp_1604001098328/work
zope.interface==6.0
